refactor(debug): extract printing functions and simplify heredoc logic
• **Extract debugging utilities** into dedicated files for better organization:
  - `print_tokens()` moved to `src/print_tokens.c`
  - `ft_prints()` moved to `src/ft_prints.c`
  - `print_parser()` moved to `src/print_parser.c`
• **Refactor heredoc processing** in `process_heredocs()` with helper functions:
  - `find_last_delimiter()` extracts delimiter selection logic
  - `read_heredoc_input()` handles user input and file writing
  - `process_heredoc_delimiters()` manages multiple heredoc processing
  - `process_fallback_heredoc()` maintains backward compatibility
• **Update header declarations** in `main_extra.h` for extracted functions
• **Add build targets** in `Makefile` for new source files
• Improves code maintainability by separating concerns and reducing function complexity
