+++ include/main_extra.h
@@ -6,7 +6,7 @@
 /*   By: lfiorell@student.42nice.fr <lfiorell>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/16 14:41:12 by jfranc            #+#    #+#             */
-/*   Updated: 2025/07/16 20:08:13 by lfiorell@st      ###   ########.fr       */
+/*   Updated: 2025/07/16 20:33:10 by lfiorell@st      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,6 +21,10 @@
 /// @param tokens Linked list of tokens to print
 void	print_tokens(t_list *tokens);
 
+/// @brief Prints command details for debugging
+/// @param cmd Pointer to the command structure to print
+void	ft_prints(t_cmd *cmd);
+
 /// @brief Prints parser state and command list for debugging
 /// @param parser Pointer to the parser structure to print
 void	print_parser(t_parser *parser);
@@ -31,5 +35,7 @@ void	print_parser(t_parser *parser);
 void	process_heredocs(t_cmd *cmds, t_list *env);
 
 t_cmd	*remove_empty_commands(t_cmd *commands);
+
+void	print_parser(t_parser *parser);
 
 #endif
