+++ src/main_extra.c
@@ -6,11 +6,12 @@
 /*   By: lfiorell@student.42nice.fr <lfiorell>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/16 14:41:12 by jfranc            #+#    #+#             */
-/*   Updated: 2025/07/16 20:23:05 by jfranc           ###   ########.fr       */
+/*   Updated: 2025/07/16 20:36:50 by lfiorell@st      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
+#include "main_extra.h"
 #include "parser.h"
 #include "reader.h"
 #include "shared.h"
@@ -19,169 +20,117 @@
 #include <stdio.h>
 #include <unistd.h>
 
-void	print_tokens(t_list *tokens)
+// Helper function to find the last delimiter in the list
+static char	*find_last_delimiter(t_list *heredoc_delimiters)
 {
-	t_list	*current;
-	t_token	*token;
+	t_list	*delim_node;
+	char	*last_delimiter;
 
-	current = tokens;
-	while (current)
+	delim_node = heredoc_delimiters;
+	last_delimiter = NULL;
+	while (delim_node)
 	{
-		token = (t_token *)current->content;
-		if (token)
+		last_delimiter = (char *)delim_node->content;
+		delim_node = delim_node->next;
+	}
+	return (last_delimiter);
+}
+
+// Helper function to read heredoc input until delimiter is found
+static void	read_heredoc_input(char *delimiter, t_file *tmpf,
+		bool write_to_file)
+{
+	char	*line;
+
+	while (true)
+	{
+		line = readline("heredoc> ");
+		if (!line)
+			break ;
+		if (ft_strncmp(line, delimiter, ft_strlen(delimiter)) == 0
+			&& ft_strlen(line) == ft_strlen(delimiter))
 		{
-			printf("Token: '%s', Type: %d\n", token->value, token->type);
+			free(line);
+			break ;
 		}
-		current = current->next;
+		if (write_to_file && tmpf)
+		{
+			ft_putstr_fd(line, tmpf->fd);
+			ft_putstr_fd("\n", tmpf->fd);
+		}
+		free(line);
 	}
 }
 
-static void	ft_prints(t_cmd *cmd)
+// Helper function to process all heredoc delimiters for a command
+static t_file	*process_heredoc_delimiters(t_list *heredoc_delimiters)
 {
-	printf("Command: %s, Arg Count: %d", cmd->args[0], cmd->argc);
-	if (cmd->is_pipe)
-		printf(" (Pipe)");
-	if (cmd->fd_infile != -1)
-		printf(", Infile: %d", cmd->fd_infile);
-	if (cmd->fd_outfile != -1)
-		printf(", Outfile: %d", cmd->fd_outfile);
-	if (cmd->redirect_in)
-		printf(", Redirect In: %s", cmd->redirect_in);
-	if (cmd->redirect_out)
-		printf(", Redirect Out: %s", cmd->redirect_out);
-	if (cmd->redirect_append)
-		printf(", Redirect Append: %s", cmd->redirect_append);
-	if (cmd->redirect_heredoc)
-		printf(", Redirect Heredoc: %s", cmd->redirect_heredoc);
-	printf("\n");
+	t_list	*delim_node;
+	char	*delimiter;
+	char	*last_delimiter;
+	t_file	*tmpf;
+
+	last_delimiter = find_last_delimiter(heredoc_delimiters);
+	tmpf = NULL;
+	delim_node = heredoc_delimiters;
+	while (delim_node)
+	{
+		delimiter = (char *)delim_node->content;
+		if (delimiter == last_delimiter)
+		{
+			tmpf = ft_opentmp(ft_openurand(), true);
+			if (!tmpf)
+			{
+				perror("heredoc");
+				break ;
+			}
+		}
+		read_heredoc_input(delimiter, tmpf, delimiter == last_delimiter);
+		delim_node = delim_node->next;
+	}
+	return (tmpf);
 }
 
-void	print_parser(t_parser *parser)
+// Helper function to handle fallback heredoc (backward compatibility)
+static t_file	*process_fallback_heredoc(char *redirect_heredoc)
 {
-	t_list	*current;
-	t_cmd	*cmd;
+	t_file	*tmpf;
 
-	if (!parser)
+	tmpf = ft_opentmp(ft_openurand(), true);
+	if (!tmpf)
 	{
-		printf("Parser is NULL\n");
-		return ;
+		perror("heredoc");
+		return (NULL);
 	}
-	printf("Parser State: %d\n", parser->state);
-	printf("Current Index: %zu\n", parser->current_index);
-	printf("Token Count: %zu\n", parser->token_count);
-	if (parser->command_list)
-	{
-		printf("Command List:\n");
-		current = parser->command_list;
-		while (current)
-		{
-			cmd = (t_cmd *)current->content;
-			if (cmd)
-				ft_prints(cmd);
-			current = current->next;
-		}
-	}
-	else
-		printf("No commands parsed.\n");
+	read_heredoc_input(redirect_heredoc, tmpf, true);
+	return (tmpf);
 }
 
 // New: process here-documents for parsed commands
 void	process_heredocs(t_cmd *cmds, t_list *env)
 {
 	size_t	iteri;
-	char	*line;
 	t_file	*tmpf;
-	t_list	*delim_node;
-	char	*delimiter;
-	char	*last_delimiter;
 
 	(void)env;
-	for (iteri = 0; cmds[iteri].args; ++iteri)
+	iteri = 0;
+	while (cmds[iteri].args)
 	{
+		tmpf = NULL;
 		if (cmds[iteri].heredoc_delimiters)
 		{
-			// Process all heredoc delimiters, but only use the last one
-			delim_node = cmds[iteri].heredoc_delimiters;
-			last_delimiter = NULL;
-			tmpf = NULL;
-			// Find the last delimiter
-			while (delim_node)
-			{
-				last_delimiter = (char *)delim_node->content;
-				delim_node = delim_node->next;
-			}
-			// Process each heredoc delimiter
-			delim_node = cmds[iteri].heredoc_delimiters;
-			while (delim_node)
-			{
-				delimiter = (char *)delim_node->content;
-				// Only open the temp file for the last delimiter
-				if (delimiter == last_delimiter)
-				{
-					tmpf = ft_opentmp(ft_openurand(), true);
-					if (!tmpf)
-					{
-						perror("heredoc");
-						break ;
-					}
-				}
-				// Always prompt for input for each heredoc
-				while (true)
-				{
-					line = readline("heredoc> ");
-					if (!line)
-						break ;
-					if (ft_strncmp(line, delimiter, ft_strlen(delimiter)) == 0
-						&& ft_strlen(line) == ft_strlen(delimiter))
-					{
-						free(line);
-						break ;
-					}
-					// Only write to file if this is the last heredoc
-					if (delimiter == last_delimiter && tmpf)
-					{
-						ft_putstr_fd(line, tmpf->fd);
-						ft_putstr_fd("\n", tmpf->fd);
-					}
-					free(line);
-				}
-				delim_node = delim_node->next;
-			}
-			// Set the file descriptor for the last heredoc
-			if (tmpf)
-			{
-				lseek(tmpf->fd, 0, SEEK_SET);
-				cmds[iteri].fd_infile = tmpf->fd;
-				free(tmpf);
-			}
+			tmpf = process_heredoc_delimiters(cmds[iteri].heredoc_delimiters);
 		}
 		else if (cmds[iteri].redirect_heredoc)
 		{
-			// Fallback for backward compatibility
-			tmpf = ft_opentmp(ft_openurand(), true);
-			if (!tmpf)
-			{
-				perror("heredoc");
-				continue ;
-			}
-			while (true)
-			{
-				line = readline("heredoc> ");
-				if (!line)
-					break ;
-				if (ft_strncmp(line, cmds[iteri].redirect_heredoc,
-						ft_strlen(cmds[iteri].redirect_heredoc)) == 0)
-				{
-					free(line);
-					break ;
-				}
-				ft_putstr_fd(line, tmpf->fd);
-				ft_putstr_fd("\n", tmpf->fd);
-				free(line);
-			}
+			tmpf = process_fallback_heredoc(cmds[iteri].redirect_heredoc);
+		}
+		if (tmpf)
+		{
 			lseek(tmpf->fd, 0, SEEK_SET);
 			cmds[iteri].fd_infile = tmpf->fd;
 			free(tmpf);
 		}
+		iteri++;
 	}
 }
