+++ src/print_parser.c
@@ -1,0 +1,43 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   print_parser.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: lfiorell@student.42nice.fr <lfiorell>      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/16 20:29:30 by lfiorell@st       #+#    #+#             */
+/*   Updated: 2025/07/16 20:29:38 by lfiorell@st      ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "parser.h"
+#include "utils.h"
+
+void	print_parser(t_parser *parser)
+{
+	t_list *current;
+	t_cmd *cmd;
+
+	if (!parser)
+	{
+		printf("Parser is NULL\n");
+		return ;
+	}
+	printf("Parser State: %d\n", parser->state);
+	printf("Current Index: %zu\n", parser->current_index);
+	printf("Token Count: %zu\n", parser->token_count);
+	if (parser->command_list)
+	{
+		printf("Command List:\n");
+		current = parser->command_list;
+		while (current)
+		{
+			cmd = (t_cmd *)current->content;
+			if (cmd)
+				ft_prints(cmd);
+			current = current->next;
+		}
+	}
+	else
+		printf("No commands parsed.\n");
+}
