+++ src/print_tokens.c
@@ -1,0 +1,33 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   print_tokens.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: lfiorell@student.42nice.fr <lfiorell>      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/07/16 20:30:00 by lfiorell@st      #+#    #+#             */
+/*   Updated: 2025/07/16 20:30:00 by lfiorell@st      ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "lexer.h"
+#include "libft.h"
+#include "main_extra.h"
+#include <stdio.h>
+
+void	print_tokens(t_list *tokens)
+{
+	t_list	*current;
+	t_token	*token;
+
+	current = tokens;
+	while (current)
+	{
+		token = (t_token *)current->content;
+		if (token)
+		{
+			printf("Token: '%s', Type: %d\n", token->value, token->type);
+		}
+		current = current->next;
+	}
+}
